rename mcp-nacos to mcp-discovery
