--- auto-configurations/spring-ai-alibaba-autoconfigure-nacos-mcp-server/src/main/resources/META-INF/spring-configuration-metadata.json
@@ -1,102 +1,0 @@
-{
-  "groups": [
-    {
-      "name": "spring.ai.alibaba.mcp.nacos",
-      "type": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "description": "Configuration properties for Nacos MCP integration."
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.registry",
-      "type": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "description": "Configuration properties for Nacos MCP server registration."
-    }
-  ],
-  "properties": [
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.server-addr",
-      "type": "java.lang.String",
-      "description": "Nacos server address.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": "localhost:8848"
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.namespace",
-      "type": "java.lang.String",
-      "description": "Nacos namespace for MCP service.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.username",
-      "type": "java.lang.String",
-      "description": "Nacos username to authenticate.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.password",
-      "type": "java.lang.String",
-      "description": "Nacos password to authenticate.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.access-key",
-      "type": "java.lang.String",
-      "description": "Nacos access key to authenticate.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.secret-key",
-      "type": "java.lang.String",
-      "description": "Nacos secret key to authenticate.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.endpoint",
-      "type": "java.lang.String",
-      "description": "Nacos server endpoint.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.NacosMcpProperties",
-      "defaultValue": ""
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.registry.enabled",
-      "type": "java.lang.Boolean",
-      "description": "Whether to register MCP server service to Nacos.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "defaultValue": false
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.registry.service-name",
-      "type": "java.lang.String",
-      "description": "Service name for MCP server service.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "defaultValue": "mcp-server"
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.registry.service-group",
-      "type": "java.lang.String",
-      "description": "Group for MCP server service.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "defaultValue": "DEFAULT_GROUP"
-    },
-    {
-      "name": "spring.ai.alibaba.mcp.nacos.registry.sse-export-context-path",
-      "type": "java.lang.String",
-      "description": "Base context path for SSE MCP server.",
-      "sourceType": "com.alibaba.cloud.ai.mcp.nacos.registry.NacosMcpRegistryProperties",
-      "defaultValue": "/mcp"
-    },
-    {
-      "name": "spring.ai.mcp.server.enabled",
-      "type": "java.lang.Boolean",
-      "description": "Whether to enable MCP server.",
-      "defaultValue": true
-    }
-  ],
-  "hints": []
-}
