+++ spring-ai-alibaba-mcp/pom.xml
@@ -16,7 +16,7 @@
 
     <modules>
         <module>spring-ai-alibaba-mcp-nacos2</module>
-        <module>spring-ai-alibaba-mcp-nacos</module>
+        <module>spring-ai-alibaba-mcp-discovery</module>
         <module>spring-ai-alibaba-mcp-nacos-common</module>
         <module>spring-ai-alibaba-mcp-gateway</module>
     </modules>
